+++ demo/demo-alipay/advert/index.tsx
@@ -1,39 +1,87 @@
-import { Component } from 'Naruse';
-
+import { Component } from "Naruse";
 
 export default class Qwer extends Component {
-
     async componentDidMount() {
-        const qq = await import('./qwer.css');
-        console.log('ccc', qq.ccc);
+        // const qq = await import('./qwer.css');
+        // console.log('ccc', qq.ccc);
     }
 
     render() {
-        console.log('render', this);
-        return <view
-        style={{
-            backgroundColor: 'red',
-            width: '100px',
-            height: '10000px',
-         }}>
-            <web-view ref='web' src='https://fuwu.taobao.com/ser/assembleParam.htm?spm=a1z13.2196529.0.0.1b1f519fmbgMhQ&tracelog=search&activityCode=&promIds=&subParams=itemCode:FW_GOODS-1827490-1,cycleNum:12,cycleUnit:2' ></web-view>
-            {new Array(100).fill(1).map((item, index) => {
-                return <view
+        return (
+            <view>
+                {this.renderRadio()}
+                {this.renderOnFirstAppear()}
+            </view>
+        );
+    }
+
+    renderOnFirstAppear() {
+        return (
+            <view
                 style={{
-                    backgroundColor: 'blue',
-                    width: '100px',
-                    height: '100px',
-                 }}
-                 onFirstAppear={() => {
-                    console.log('onFristAppear', index);
+                    backgroundColor: "red",
+                    width: "100px",
+                    height: "10000px",
                 }}
-                onClick={() => {
-                    console.log('onClick', index);
-                }}
-                 >
-                    {index}
-                </view>;
+            >
+                {new Array(100).fill(1).map((item, index) => {
+                    return (
+                        <view
+                            style={{
+                                backgroundColor: "blue",
+                                width: "100px",
+                                height: "100px",
+                            }}
+                            onFirstAppear={() => {
+                                console.log("onFristAppear", index);
+                            }}
+                            onClick={() => {
+                                console.log("onClick", index);
+                            }}
+                        >
+                            {index}
+                        </view>
+                    );
                 })}
-        </view>;
+            </view>
+        );
+    }
+
+    renderRadio() {
+        const { radioValue = "b" } = this.state;
+        return (
+            <view>
+                当前选项值：{radioValue}
+                <radio-group onChange={(evl) => console.log(evl)}>
+                    <radio value={"a"}>a选项</radio>
+                    <radio value={"b"}>b选项</radio>
+                </radio-group>
+                <checkbox-group onChange={(evl) => console.log(evl)} disabled>
+                    <checkbox value={"aa"}>aa选项</checkbox>
+                    <checkbox value={"bb"} checked>
+                        bb选项
+                    </checkbox>
+                    <checkbox value={"cc"}>cc选项</checkbox>
+                </checkbox-group>
+                
+                <switch onChange={(evl) => console.log(evl)} />
+                <view style={{ width: "80%", margin: "0 auto" }}>
+                    <slider
+                        min={-20}
+                        max={20}
+                        // step={5}
+                        // value={50}
+                        // showValue
+                        // activeColor="#000"
+                        // backgroundColor="#fff"
+                        // trackSize={15}
+                        // handleSize={50}
+                        // handleColor="#eee"
+                        onChanging={(evl) => console.log(evl)}
+                        onChange={(evl) => console.log(evl)}
+                    />
+                </view>
+            </view>
+        );
     }
 }