+++ package/fs-aw/src/index.ts
@@ -1,0 +1,3 @@
+export * as webFs from './index.web';
+export * as miniFs from './index.miniapp';
+export * as weexFs from './index.weex';