+++ package/naruse-webpack-runner/src/mxml/data.ts
@@ -1,7 +1,4 @@
 const speicalConfig = {
-    attributes: {
-        'class': 'className',
-    },
     // 基础属性
     baseAttributes: {
         'style': 'style',
@@ -103,6 +100,119 @@ const scrollView = {
         ...baseElementEvents,
     ],
     nest: true,
+}
+
+/**
+ * 单选组
+ */
+const radioGroup = {
+    templateName: 'radio-group',
+    reflectTagName: 'radio-group',
+    attributes: [
+        'name',
+    ],
+    events: [
+        'onChange',
+        ...baseElementEvents,
+    ],
+    nest: true,
+}
+
+/**
+ * 原生 单选
+ */
+export const nativeRadio = {
+    templateName: 'native-radio',
+    reflectTagName: 'radio',
+    attributes: [
+        'value',
+        'checked',
+        'disabled',
+        'color',
+    ],
+    events: [
+        ...baseElementEvents,
+    ],
+}
+
+/**
+ * 复选框组
+ */
+const checkboxGroup = {
+    templateName: 'checkbox-group',
+    reflectTagName: 'checkbox-group',
+    attributes: [
+        'name',
+    ],
+    events: [
+        'onChange',
+        ...baseElementEvents,
+    ],
+    nest: true,
+}
+
+/**
+ * 原生 复选框
+ */
+export const nativeCheckbox = {
+    templateName: 'native-checkbox',
+    reflectTagName: 'checkbox',
+    attributes: [
+        'value',
+        'checked',
+        'disabled',
+        'color',
+    ],
+    events: [
+        'onChange',
+        ...baseElementEvents,
+    ],
+}
+
+/**
+ * 单选开关
+ */
+export const nativeSwitch = {
+    templateName: 'switch',
+    reflectTagName: 'switch',
+    attributes: [
+        'name',
+        'checked',
+        'disabled',
+        'color',
+        'controlled',
+    ],
+    events: [
+        'onChange',
+        ...baseElementEvents,
+    ],
+}
+
+/**
+ * 滑动选择器
+ */
+export const slider = {
+    templateName: 'slider',
+    reflectTagName: 'slider',
+    attributes: [
+        'name',
+        'min',
+        'max',
+        'step',
+        'disabled',
+        'value',
+        'show-value',
+        'active-color',
+        'background-color',
+        'track-size',
+        'handle-size',
+        'handle-color',
+    ],
+    events: [
+        'onChange',
+        'onChanging',
+        ...baseElementEvents,
+    ],
 }
 
 const text = {
@@ -272,6 +382,12 @@ const templateList = [
     pureView,
     pureText,
     webView,
+    nativeRadio,
+    radioGroup,
+    checkboxGroup,
+    nativeCheckbox,
+    nativeSwitch,
+    slider,
 ]